Module Setup screen: push pins, flags and valve tuning to a module
The panel behind Rate Control -> Module setup edits the settings a rate
module keeps in its own EEPROM, and sends them on demand. Until now
those could only be written by the WinForms RateController app, which no
longer exists in the tractor - the module's own web page is just WiFi
and a master button.

Saved per TOOL (modules-<tool>.json), because the module is bolted to
that implement and its tuning must never leak to another. Nothing can
read a module's config back, so this file becomes the master record: a
reset or replaced board is recommissioned with 'Send everything'.

Sends are explicit, never on the periodic tick, since they make the
module save and pin changes restart it. ID assignment is guarded twice -
refused while more than one module is answering, then a confirm - because
the module adopts it unconditionally and two boards sharing an ID both
stop working.

Pin arrays persist as JSON numbers rather than byte[], which
System.Text.Json would write as base64 and make the record unreadable.

Verified against the running app: panel reads /api/ratemodules, edits
persist per tool, and the seat gate applies. Suite 1061/0.

+++ Shared/AgOpenWeb.Services/RateControl/RcPgn.cs
@@ -113,10 +113,13 @@ public sealed class RcModuleConfig
     public byte OnboardRelayType { get; set; }
     /// <summary>Same encoding as <see cref="OnboardRelayType"/>.</summary>
     public byte RemoteRelayType { get; set; }
-    /// <summary>Flow/dir/PWM pins for sensor 0 and sensor 1 (6 bytes, in that order).</summary>
-    public byte[] SensorPins { get; set; } = new byte[6];
+    // int[] rather than byte[] purely so these persist as readable JSON arrays -
+    // System.Text.Json writes byte[] as base64, which would make the saved setup
+    // (the master record of an implement) opaque and un-editable by hand.
+    /// <summary>Flow/dir/PWM pins for sensor 0 and sensor 1 (6 values, in that order).</summary>
+    public int[] SensorPins { get; set; } = new int[6];
     /// <summary>Relay output pins 0-15.</summary>
-    public byte[] RelayPins { get; set; } = new byte[16];
+    public int[] RelayPins { get; set; } = new int[16];
     public byte WorkPin { get; set; }
     public byte PressurePin { get; set; }
 }
@@ -297,10 +300,10 @@ public static class RcPgn
         d[4] = cmd;
         d[5] = c.OnboardRelayType;
         d[6] = c.RemoteRelayType;
-        var sp = c.SensorPins ?? Array.Empty<byte>();
-        for (int i = 0; i < 6 && i < sp.Length; i++) d[7 + i] = sp[i];
-        var rp = c.RelayPins ?? Array.Empty<byte>();
-        for (int i = 0; i < 16 && i < rp.Length; i++) d[13 + i] = rp[i];
+        var sp = c.SensorPins ?? Array.Empty<int>();
+        for (int i = 0; i < 6 && i < sp.Length; i++) d[7 + i] = (byte)sp[i];
+        var rp = c.RelayPins ?? Array.Empty<int>();
+        for (int i = 0; i < 16 && i < rp.Length; i++) d[13 + i] = (byte)rp[i];
         d[29] = c.WorkPin;
         d[30] = c.PressurePin;
         d[31] = 0;     // spare
